feat: add `resetDetector()` method to be able to start a fresh detection
Due to their need to detect changes or repetitions between two frames, both detector objects keep some values related to previous frame(s) processing in their internal state.

To be able to start a new detection on a completly new frame (meaning a frame which is not the direct successor of the last frame handled by the detectors), we add this `resetDetector()` method which reset these internal state values to their starting default.