fix: outputs were automatically started but shoudn't
Before this change, we were setting the encoder outputs BEFORE calling `start_encoder()`, which would start the encoder AND the outputs.

This was problematic because in case of the `streamingOutput`, it would attempt to connect to the remote rtmp server on start even though we didn't want to connect yet.

Setting the encoder outputs AFTER calling `start_encoder()` allows to start the encoder only and leave the outputs to be started later on when we need to.