chore: adapt detection and recording config values for best results
The previous values were maybe a little too strict and thus either were too slow to catch or did not catch at all rapid movements like leaving with a car, passing by in bicycle, this sort of things.
These new values are a test to see how it performs, keeping in mind we always have to keep a balance between fast detection and false positives.