feat: add recording phase to record the detected object
The recording phase is the whole period of time while we are recording the detection of the object.
It is the very first phase which starts as soon as the object is detected and lasts until an object (be it the object originally detected or not) is detected for the last time + 8 seconds OR until the max duration is reached (currently set at 120 seconds).