feat: remove waiting phase
The waiting phase was originally thought to have the following benefits:
- let time to the CPU to cool down by having no computation to do
- avoid firing too many detection notifications to the user

However I saw it more like a bad for a good because it could in certain cases prevent detections that would be of interest.

Now that I am more confident that a fan and a heatsink can efficiently cool down the CPU while running under heavy loads, and realizing that the checking phase already reduces the detection notifications to what the user only needs, we can remove the waiting phase which brings no major benefits anymore.
This also has the advantage to simplify the code logic and thus reduce the error/bug surface.