feat: speed up video start by moving metadata to the beginning
The FFMPEG  `-movflags faststart` allows to put the metadata needed by the browser to start playing to the beginning of the video.
Hence the browser doesn't need to download the full file or the end of the video just to get the video metadata.

Adding this flag has significantly reduced the time needed for the browser to display the first video frame (the automatic "poster") and start playing.