chore: add `websockets` `v12.0` to dependencies
When I tried to set up the project, a too old (v8.1) `websockets` package was already installed on the system and used. It led to problems when trying to run the project because the old package did not contain necessary classes.
Therefore we specify this working version of the package to be installed and used.